Implemented “Discover More NFTs” and “Top Creators” sections on landing page
