+++ src/app/layout.tsx
@@ -1,39 +1,43 @@
-import type { Metadata } from 'next';
-import { Geist, Geist_Mono, Inter, Work_Sans, Jersey_10 } from 'next/font/google';
-import './globals.css';
-import NavBar from '@/components/navbar';
+import type { Metadata } from "next";
+import {
+  Geist,
+  Geist_Mono,
+  Inter,
+  Work_Sans,
+  Jersey_10,
+} from "next/font/google";
+import "./globals.css";
+import NavBar from "@/components/home/layout/navbar";
 
 const geistSans = Geist({
-  variable: '--font-geist-sans',
-  subsets: ['latin'],
+  variable: "--font-geist-sans",
+  subsets: ["latin"],
 });
 
 const geistMono = Geist_Mono({
-  variable: '--font-geist-mono',
-  subsets: ['latin'],
+  variable: "--font-geist-mono",
+  subsets: ["latin"],
 });
 
 const inter = Inter({
-  variable: '--font-inter',
-  subsets: ['latin'],
+  variable: "--font-inter",
+  subsets: ["latin"],
 });
 
-
 const workSans = Work_Sans({
-  variable: '--font-work-sans',
-  subsets: ['latin'],
+  variable: "--font-work-sans",
+  subsets: ["latin"],
 });
 
 const Jeysey10 = Jersey_10({
-  variable: '--font-jersey-10',
-  subsets: ['latin'],
-  weight: '400'
+  variable: "--font-jersey-10",
+  subsets: ["latin"],
+  weight: "400",
 });
 
-
 export const metadata: Metadata = {
-  title: 'Create Next App',
-  description: 'Generated by create next app',
+  title: "Create Next App",
+  description: "Generated by create next app",
 };
 
 export default function RootLayout({
@@ -43,11 +47,11 @@ export default function RootLayout({
 }>) {
   return (
     <html lang="en">
-      <head>
-      </head>
+      <head></head>
       <body
         suppressHydrationWarning
-        className={`${geistSans.variable} ${workSans.variable} ${inter.variable} ${Jeysey10.variable} ${geistMono.variable} antialiased`}>
+        className={`${geistSans.variable} ${workSans.variable} ${inter.variable} ${Jeysey10.variable} ${geistMono.variable} antialiased`}
+      >
         <NavBar />
         {children}
       </body>
