+++ src/app/page.tsx
@@ -1,9 +1,13 @@
-import { HeroSection } from "@/components/home/HeroSection";
+import { HeroSection } from "@/components/home/sections/HeroSection";
+import { DiscoverNFTsSection } from "@/components/home/sections/DiscoverNFTsSection";
+import { TopRatedArtist } from "@/components/home/sections/TopRatedArtist";
 
 export default function Home() {
   return (
     <>
     <HeroSection />
+    <DiscoverNFTsSection />
+    <TopRatedArtist />
     </>
   );
 }
